requirements.txt
requests
pillow
RPi.GPIO
spidev
